Refactor: Split aggregate_resolver_sumcount into two files
This change splits the aggregate_resolver_sumcount.cpp file into two separate files: aggregate_resolver_sumcount_only.cpp and aggregate_resolver_distinct.cpp. This improves compilation performance by reducing dependencies and organizing code more logically.
